# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/api/version-info": {
            "get": {
                "tags": [
                    "Server Info"
                ],
                "summary": "Get Server Version Info",
                "description": "Get Server Version Info",
                "operationId": "versionInfo",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Server Version Information Retrieved Successfully"
                                        },
                                        "serverVersion": {
                                            "type": "string",
                                            "example": "0.0.1"
                                        },
                                        "lastReleaseDate": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2023-08-22"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Something Went Wrong."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}